Add asset commands to readme
As a complete Wagtail noob on support I needed to run the NHSX site
locally and found all the test failed due to a missing Apple touch icon.

Speaking with a developer, running these commands resolved the issue, so
I felt it was worth adding them to the readme for other folks.